move to capapcity config
