+++ electricitymap/contrib/capacity_parsers/EIA.py
@@ -6,8 +6,8 @@ from typing import Any
 import pandas as pd
 from requests import Response, Session
 
-from electricitymap.contrib.capacity_parsers import CAPACITY_PARSER_SOURCE_TO_ZONES
 from electricitymap.contrib.config import ZoneKey
+from electricitymap.contrib.config.capacity import CAPACITY_PARSER_SOURCE_TO_ZONES
 from electricitymap.contrib.config.constants import PRODUCTION_MODES
 from parsers.EIA import REGIONS
 from parsers.lib.utils import get_token
