+++ electricitymap/contrib/capacity_parsers/IRENA.py
@@ -6,8 +6,8 @@ from typing import Any
 import pycountry
 from requests import Response, Session
 
-from electricitymap.contrib.capacity_parsers import CAPACITY_PARSER_SOURCE_TO_ZONES
 from electricitymap.contrib.config import ZoneKey
+from electricitymap.contrib.config.capacity import CAPACITY_PARSER_SOURCE_TO_ZONES
 
 """The data is downloaded from the IRENA API. """
 logger = getLogger(__name__)
